Swerve drive is semi-working! (rounding error)

+++ RobotC/Headers/structs.h
@@ -8,7 +8,7 @@
 typedef struct motorData {
 	bool isReversed;
 	int angleOffset;
-	int fineTuneFactor;
+	float fineTuneFactor;
 	int power;
 } motorData;
